--- frontend/src/components/Navbar.module.css
@@ -1,3 +1,0 @@
-.container {
-    padding: 10px;
-}
