+++ frontend/src/hooks/useCategoryLocation.ts
@@ -1,0 +1,6 @@
+import { useLocation } from 'react-router-dom';
+
+export const useCategoryLocation = () => {
+  const location = useLocation()
+  return location.pathname.split('/')[1]
+};
